[Feat] 회원 정보 조회, 수정, 삭제 기능 추가

+++ src/main/java/org/team10/washcode/entity/LaundryShop.java
@@ -23,8 +23,8 @@ public class LaundryShop {
     private String address;
     private String non_operating_days;
 
-    private int latitude;
-    private int longitude;
+    private Double latitude;
+    private Double longitude;
 
     private Timestamp created_at;
 }
